Remove unnecessary ldr from boot wrapper
We currently do an ldr from GICC_CTLR to w0, then immediately overwrite
w0 with a mov. Reading the GICC_CTLR has no effect on the state of the
GIC, so there's no reason to do the ldr. It's also inconsistent with the
way we set the GICD_CTLR.

Fix this.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>