Allow support of multiple assembly files
To support more complex functionality, it would be nice if we could
separate bits of the boot code into separate files.

This patch refactors the Makefile, allowing us to add more source files
later without having to add a new rule for each file. The defines we
pass to each object are also factored into separate lines for easier
modification in future.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>