gitignore: ignore build files
Currently. we don't ignore several files produced during the build in
the .gitignore, which can make the output of git status very noisy.

Let's ignore all of the external dependencies we might add to the source
directory (dtc, *.dts{i,}, *.cpio.gz), and all of the files we'll
produce during the build (*.o, *.axf). Ignoring the general case of all
of these files should produce less churn in .gitignore in future, and we
can always add special exemptions if required.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>