bootwrapper: improve UART initialisation
The PL011 spec requires that baud rate settings should be done by
writing both xBRD registers followed by a write to LCR_H (which we
skipped so far completely).[1]
So include the FBRD register write as well and set up the LCR_H
register to initialise the UART to 8N1.
Also skip the automatic RTS/CTS enablement, as this may not be
used by the peer (and the generic UART does not have it at all).

[1] ARM PL011 TRM, 3.3.7 (top of page 3-14)

Signed-off-by: Andre Przywara <andre.przywara@arm.com>