configure: fix file detection when cross-compiling
The autotools documentation states that AC_CHECK_FILE cannot be used when
cross-compiling [1], because it's meant to check files in the target
system, not on the build host. When just giving --host on the configure
command line, the script detects cross compilation rather late; and as the
file test just happens to execute earlier, this works anyway.
However if one gives both --host and --build, cross compilation is
detected very early and ./configure complains:

checking for /src/linux-arm64... configure: error: cannot check for file existence when cross compiling

So replace the checkfile macro usage with a simple "test -f" call (which
is the recommended way of checking for files on the build host) and output
proper error messages.

[1] https://www.gnu.org/software/autoconf/manual/autoconf.html#Files

Signed-off-by: Andre Przywara <andre.przywara@arm.com>
[Mark: simplified error messages]
Signed-off-by: Mark Rutland <mark.rutland@arm.com>