Dynamically determine the set of CPUs
Currently we hard-code the set of CPUs we expect, and we have some
strong expectations on the formatting of nodes.

As we can configure models with differing sets of CPUs, we added the
with-cpu-ids configure option to override this assumption, though in
practice it turns out this is very fragile.

Instead, we can parse the DTB to discover the set of CPU nodes (and
hence the set of CPU IDs, and the number of CPUs). This is far more
robust.

This patch changes the bootwrapper to do this, removing the newly
redundant --with-cpu-ids configure option.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>